requirements.txt
sqlalchemy
pymysql
python-dotenv
dash
dash-bootstrap-components
pandas
numpy
plotly
